requirements.txt
ffmpeg-python
pillow
python-dotenv
requests
urllib3
